Update to roll dev out to production

+++ CCIA.SQL/dbo/Tables/cert_rad.sql
@@ -1,7 +1,9 @@
 ﻿CREATE TABLE [dbo].[cert_rad] (
-    [cert_num]  INT      NOT NULL,
-    [cert_year] SMALLINT NOT NULL,
-    [rad]       SMALLINT NOT NULL,
+    [cert_num]  INT NOT NULL,
+    [cert_year] INT NOT NULL,
+    [rad]       INT NOT NULL,
     CONSTRAINT [PK_cert_rad] PRIMARY KEY NONCLUSTERED ([cert_num] ASC, [cert_year] ASC, [rad] ASC)
 );
+
+
